forms: Typecast values to correct Type
Before this change every value from the model was casted to str()
before passing it to the Django form field.
This works fine for text- and number bseed formats.
However for BooleanFields and CheckBoxes this breaks.

This change now checks the type of the targt and casts the values to the
correct type.

Signed-off-by: Chris Fiege <cfi@pengutronix.de>